requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
qdrant-client==1.12.1
numpy==2.1.3
